Refactored directory
