+++ lib/constructs/LangfuseEcrStackProps.ts
