+++ lib/constructs/langfuse-ecr-fargate-deployment-cdk-stack.ts
@@ -85,7 +85,7 @@ export class CdkFargateWithVpcDeploymentStack extends cdk.NestedStack {
                     "logs:CreateLogStream",
                     "logs:PutLogEvents"
                 ],
-                resources: ["*"],
+                resources: [props.ecrRepository.repositoryArn],
             }
         ));
